Adjust the ipmi command test. Rename the test. The verification of the emu data will be tested in another case.
